+++ src/js/dropdowns.js
@@ -10,8 +10,8 @@ var jqLite = require('./lib/jqLite.js'),
     util = require('./lib/util.js'),
     attrKey = 'data-mui-toggle',
     attrSelector = '[data-mui-toggle="dropdown"]',
-    openClass = 'mui-is-open',
-    menuClass = 'mui-dropdown-menu';
+    openClass = 'mui--is-open',
+    menuClass = 'mui-dropdown__menu';
 
 
 /**