+++ src/js/ripple.js
@@ -9,8 +9,7 @@
 var jqLite = require('./lib/jqLite.js'),
     util = require('./lib/util.js'),
     btnClass = 'mui-btn',
-    btnStyleKey = 'data-mui-style',
-    btnFloatingClass = 'mui-btn-floating',
+    btnFABClass = 'mui-btn--fab',
     rippleClass = 'mui-ripple-effect',
     animationName = 'mui-btn-inserted';
 
@@ -66,13 +65,8 @@ function eventHandler(ev) {
       radius;
 
   // get height
-  // TODO: remove class check
-  if (jqLite.hasClass(buttonEl, btnFloatingClass) || 
-      buttonEl.getAttribute(btnStyleKey) === 'fab') {
-    diameter = offset.height / 2;
-  } else {
-    diameter = offset.height;
-  }
+  if (jqLite.hasClass(buttonEl, btnFABClass)) diameter = offset.height / 2;
+  else diameter = offset.height;
 
   radius = diameter / 2;