+++ src/react/buttons.jsx
@@ -8,11 +8,8 @@
 var util = require('../js/lib/util.js'),
     Ripple = require('./ripple.jsx');
 
-var buttonClass = 'mui-btn',
-    flatClass = buttonClass + '-flat',
-    raisedClass = buttonClass + '-raised',
-    largeClass = buttonClass + '-lg',
-    floatingClass = buttonClass + '-floating';
+var btnClass = 'mui-btn',
+    btnAttrs = {style: 1, color: 1, size: 1};
 
 
 /**
@@ -23,23 +20,25 @@ var Button = React.createClass({
   mixins: [Ripple],
   getDefaultProps: function() {
     return {
-      type: 'default',  // one of default, primary, danger or accent
+      style: 'default', // default|flat|raised|fab
+      color: 'default', // default|primary|danger|dark|accent
+      size: 'default', // default|small|large
       disabled: false
     };
   },
   render: function() {
-    var cs = {};
-
-    cs[buttonClass] = true;
-    cs[buttonClass + '-' + this.props.type] = true;
-    cs[flatClass] = this.props.flat;
-    cs[raisedClass] = this.props.raised;
-    cs[largeClass] = this.props.large;
-    cs = util.classNames(cs);
+    var cls = btnClass,
+        k,
+        v;
+    
+    for (k in btnAttrs) {
+      v = this.props[k];
+      if (v !== 'default') cls += ' ' + btnClass + '--' + v;
+    }
 
     return (
       <button
-        className={ cs }
+        className={ cls }
         disabled={ this.props.disabled }
         onMouseDown={ this.ripple }
         onTouchStart={ this.ripple }
@@ -53,43 +52,7 @@ var Button = React.createClass({
 });
 
 
-/**
- * Round button constructor
- * @class
- */
-var RoundButton = React.createClass({
-  mixins: [Ripple],
-  getDefaultProps: function() {
-    return {
-      floating: true
-    };
-  },
-  render: function() {
-    var cs = {};
-
-    cs[buttonClass] = true;
-    cs[floatingClass] = true;
-    cs[floatingClass + '-mini'] = this.props.mini;
-    cs = util.classNames(cs);
-
-    return (
-      <button
-        className={ cs }
-        disabled={ this.props.disabled }
-        onMouseDown={ this.ripple }
-        onTouchStart={ this.ripple }
-        onClick={ this.props.onClick }
-      >
-        { this.props.children }
-        { this.state.ripples && this.renderRipples() }
-      </button>
-    );
-  }
-})
-
-
 /** Define module API */
 module.exports = {
-  Button: Button,
-  RoundButton: RoundButton
+  Button: Button
 };